auth_api/requirements.txt
redis==4.4.2
pydantic==1.9.0
psycopg2==2.9.3
python-dotenv==0.21.0
Flask==2.2.2
gevent==22.10.2
SQLAlchemy==1.4.15
flask_sqlalchemy==2.5.1
flask-swagger-ui==4.11.1
Flask-JWT-Extended==4.4.4
pytest==7.2.1
requests==2.28.2
alembic==1.9.2
opentelemetry-api==1.10.0
opentelemetry-sdk==1.10.0
opentelemetry-instrumentation-flask==0.29b1
opentelemetry-exporter-jaeger==1.10.0
user-agents==2.2.0
Flask-Limiter==3.1.0
rauth==0.7.3
